fixed a few bugs that prevented quest data from loading
