+++ index.html
@@ -26,6 +26,6 @@
 			</label>
 			<button>TINY BUTTON!!</button>
 		</form>
-		<script type="module" src="./home/home.js"></script>
+		<script type="module" src="./app.js"></script>
 	</body>
 </html>
