+++ map/create-quest-link.js
@@ -3,7 +3,7 @@
 export function createQuest(quest) {
     const link = document.createElement('a');
     link.classList.add('quest');
-    link.href = '../quest/?id=' + quest.id;
+    link.href = `../quest/?id=${quest.id}`;
     // link.style.top = quest.map.top;
     // link.style.left = quest.map.left;
     link.textContent = quest.title;
