+++ quest/index.html
@@ -13,15 +13,15 @@
 
 		<title>Quest Index</title>
 	</head>
-	<body>
-<header id ="quest-title"></header>
+	<body><div id="user-header"></div>
+<header id ="quest-title"></div></header>
 
 
 <!-- inject quest.img-->
 <span id="quest-img"></span>
 
 <!-- inject quests.description -->
-<span id = "quest-description"></span>
+<span id = "quest-description">Quest text</span>
 
 <!-- form id = "choice-form" -->
 <form id = "choice-form">
@@ -35,7 +35,7 @@
 
 <!-- section id = results-description, class = "hidden" in css? >>> hidden until submit click -->
 <section id="results-description" class="hidden">
-<!-- <!-- inject results data based on user choice -- -->
+<!--  inject results data based on user choice -- -->
 </section>
